importccl: move sst writer proc to own file
Release note: none.
